platform.py: change manage_hotspot to switch to hotspot mode if not connected to wifi
Previously, you could "brick" the pi by connecting it to WiFi, then moving it away
from that WiFi connection. It would be configured to not run in hotspot mode, and
there's no monitoring. This commit adds the monitoring check: if the pi doesn't seem
to be connected to WiFi (multiple checks over 20 seconds), it will switch over
to hotspot mode.